fix(exports): fix csv/excel exports by removing workers
