+++ src/renderer/exporters/csv/csv-exporter-types.ts
@@ -1,0 +1,19 @@
+import type { HashesMap } from "@common/utils/hashes-types";
+
+import type {
+  AliasMap,
+  CommentsMap,
+  FilesAndFoldersMap,
+} from "../../reducers/files-and-folders/files-and-folders-types";
+import type { FilesAndFoldersMetadataMap } from "../../reducers/files-and-folders-metadata/files-and-folders-metadata-types";
+import type { TagMap } from "../../reducers/tags/tags-types";
+
+export interface CsvExporterData {
+  aliases: AliasMap;
+  comments: CommentsMap;
+  elementsToDelete: string[];
+  filesAndFolders: FilesAndFoldersMap;
+  filesAndFoldersMetadata: FilesAndFoldersMetadataMap;
+  hashes?: HashesMap;
+  tags: TagMap;
+}
