+++ src/renderer/exporters/csv/csv-exporter.controller.ts
@@ -1,3 +1,4 @@
+import { arrayToCsv } from "@common/utils/csv";
 import type { HashesMap } from "@common/utils/hashes-types";
 import type { Observable } from "rxjs";
 
@@ -8,13 +9,8 @@ import type {
 } from "../../reducers/files-and-folders/files-and-folders-types";
 import type { FilesAndFoldersMetadataMap } from "../../reducers/files-and-folders-metadata/files-and-folders-metadata-types";
 import type { TagMap } from "../../reducers/tags/tags-types";
-import { translations } from "../../translations/translations";
-import { createAsyncWorkerForChildProcessControllerFactory } from "../../utils/async-worker/child-process";
-import { backgroundWorkerProcess$ } from "../../utils/batch-process";
-import type {
-  ErrorMessage,
-  ResultMessage,
-} from "../../utils/batch-process/types";
+import { generateArrayExport$ } from "../../utils/array-export/array-export";
+import type { ResultMessage } from "../../utils/batch-process/types";
 
 export interface GenerateCsvExportOptions {
   aliases: AliasMap;
@@ -33,12 +29,4 @@ export interface GenerateCsvExportOptions {
  */
 export const generateCsvExport$ = (
   data: GenerateCsvExportOptions
-): Observable<ErrorMessage | ResultMessage> => {
-  const { language } = translations;
-  return backgroundWorkerProcess$(
-    { ...data, language },
-    createAsyncWorkerForChildProcessControllerFactory(
-      "exporters/csv/csv-exporter.fork.ts"
-    )
-  );
-};
+): Observable<ResultMessage> => generateArrayExport$(data, arrayToCsv);
