+++ src/renderer/utils/array-export/array-export.ts
@@ -1,11 +1,11 @@
 import type { HashesMap } from "@common/utils/hashes-types";
 import type { TFunction } from "i18next";
-import _ from "lodash";
+import _, { flatten } from "lodash";
 import { compose, cond, defaults, isObject, prop, stubTrue } from "lodash/fp";
-import type { Observable } from "rxjs";
-import { concat, from, interval } from "rxjs";
-import { map, take } from "rxjs/operators";
+import { concat, from, interval, Observable } from "rxjs";
+import { map, take, tap, toArray } from "rxjs/operators";
 
+import type { GenerateCsvExportOptions } from "../../exporters/csv/csv-exporter.controller";
 import { ROOT_FF_ID } from "../../reducers/files-and-folders/files-and-folders-selectors";
 import type {
   AliasMap,
@@ -14,6 +14,9 @@ import type {
 } from "../../reducers/files-and-folders/files-and-folders-types";
 import type { FilesAndFoldersMetadataMap } from "../../reducers/files-and-folders-metadata/files-and-folders-metadata-types";
 import type { TagMap } from "../../reducers/tags/tags-types";
+import { translations } from "../../translations/translations";
+import type { ResultMessage } from "../batch-process/types";
+import { MessageTypes } from "../batch-process/types";
 import { getAllChildren } from "../file-and-folders";
 import type { CellConfig } from "./make-array-export-config";
 import { makeRowConfig } from "./make-array-export-config";
@@ -161,3 +164,35 @@ export const exportToCsv: (input: CsvExporterData) => Observable<string[][]> =
     maybeRemoveDuplicates,
     addRowConfig
   );
+
+export const generateArrayExport$ = (
+  data: GenerateCsvExportOptions,
+  outputFormatter: (matrix: string[][]) => string
+): Observable<ResultMessage> => {
+  return new Observable<ResultMessage>((subscriber) => {
+    void exportToCsv({
+      ...data,
+      elementsToDelete: data.elementsToDelete ?? [],
+      translator: translations.t.bind(translations),
+    })
+      .pipe(
+        tap((row: string[][]) => {
+          subscriber.next({
+            result: row.length,
+            type: MessageTypes.RESULT,
+          });
+        }),
+        toArray()
+      )
+      .toPromise()
+      .then(flatten)
+      .then((array) => {
+        subscriber.next({
+          result: outputFormatter(array),
+          type: MessageTypes.RESULT,
+        });
+
+        subscriber.complete();
+      });
+  });
+};
